[feat] add app file
