+++ CRA_TeamProject_SSD/CRA_TeamProject_SSD.vcxproj
@@ -129,6 +129,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="TestApplication.cpp" />
     <ClCompile Include="TestShell.cpp" />
     <ClCompile Include="VirtualSSD.cpp" />
   </ItemGroup>
